backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
google-generativeai==0.3.2
python-multipart==0.0.6
Pillow==10.2.0
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
sse-starlette==1.8.2
